More common split/cleanup
