+++ src/ckzg.c
@@ -15,11 +15,12 @@
  */
 
 #include "common/alloc.c"
+#include "common/blob.c"
 #include "common/bytes.c"
-#include "common/common.c"
 #include "common/fr.c"
 #include "common/g1.c"
 #include "common/g2.c"
+#include "common/helpers.c"
 #include "common/lincomb.c"
 #include "common/ret.c"
 #include "common/settings.c"
