+++ src/common/blob.h
@@ -16,6 +16,9 @@
 
 #pragma once
 
+#include "fr.h"
+#include "ret.h"
+
 #include <inttypes.h> /* For uint*_t */
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -42,3 +45,17 @@
 typedef struct {
     uint8_t bytes[BYTES_PER_BLOB];
 } Blob;
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+// Public Functions
+////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+C_KZG_RET blob_to_polynomial(fr_t *p, const Blob *blob);
+
+#ifdef __cplusplus
+}
+#endif
