+++ src/common/bytes.h
@@ -49,6 +49,7 @@ void bytes_from_bls_field(Bytes32 *out, const fr_t *in);
 C_KZG_RET bytes_to_bls_field(fr_t *out, const Bytes32 *b);
 C_KZG_RET bytes_to_kzg_commitment(g1_t *out, const Bytes48 *b);
 C_KZG_RET bytes_to_kzg_proof(g1_t *out, const Bytes48 *b);
+void hash_to_bls_field(fr_t *out, const Bytes32 *b);
 
 #ifdef __cplusplus
 }
