+++ src/eip7594/api.c
@@ -18,6 +18,7 @@
 
 #include "common/alloc.h"
 #include "common/fr.h"
+#include "common/helpers.h"
 #include "constants.h"
 #include "fft.h"
