+++ src/eip7594/api.h
@@ -16,6 +16,7 @@
 
 #pragma once
 
+#include "common/blob.h"
 #include "common/constants.h"
 #include "common/ret.h"
 #include "common/settings.h"
