+++ src/eip7594/constants.h
@@ -17,7 +17,6 @@
 #pragma once
 
 #include "blst.h"
-#include "common/common.h"
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 // Macros
